Move post appbar in posts widget package
